requirements.txt
easyocr==1.7.2
fastapi==0.116.1
gradio==5.43.1
imutils==0.5.4
keras==3.10.0
matplotlib==3.10.5
numpy==1.24.3
onnxruntime==1.22.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
paddle
paddleocr
Pillow==11.3.0
pydantic==1.10.22
PyYAML==6.0.2
Requests==2.32.5
scikit_learn==1.7.1
scipy==1.16.1
skimage==0.0
ultralytics==8.3.167